LayoutTests/http/tests/security/private-browsing-http-auth-expected.txt

http://127.0.0.1:8000/security/resources/basic-auth.php?username=webkit&password=rocks - didReceiveAuthenticationChallenge - Responding with webkit:rocks
This test makes sure that auth credentials cached during a private browsing session do not leak out after private browsing is disabled.

--------
Frame: '<!--frame1-->'
--------
Authenticated as user: webkit password: rocks

--------
Frame: '<!--frame2-->'
--------
Resource loaded with HTTP authentication username '' and password ''
